feat(files-grid): add recently added files
