+++ src/components/files/views/grid/ImagesGrid.tsx
@@ -34,14 +34,13 @@ import { useFilesContext } from "@/context/FilesContext";
 import { ContextFile } from "@/context/FilesContext";
 import SelectingBar from "./SelectingBar";
 import FolderDescription from "@/components/folders/views/grid/FolderDescription";
+import RecentlyAdded from "./RecentlyAdded";
 
 export const ImagesGrid = ({ sortState }: { sortState: ImagesSortMethod }) => {
   const { user } = useSession();
   const { folder, isShared } = useFolderContext();
   const { files, setFiles } = useFilesContext();
 
-  const t = useTranslations("images");
-  const recentRef = useRef<HTMLDivElement>(null);
   const [carouselOpen, setCarouselOpen] = useState<boolean>(false);
   const [startIndex, setStartIndex] = useState(0);
 
@@ -388,52 +387,8 @@ export const ImagesGrid = ({ sortState }: { sortState: ImagesSortMethod }) => {
     ));
   };
 
-  const newFiles = useMemo(() => {
-    return files.filter((file) => {
-      const now = new Date();
-      const fileDate = new Date(file.createdAt);
-      const diffTime = Math.abs(now.getTime() - fileDate.getTime());
-      const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-      return diffDays <= 3;
-    });
-  }, [files]);
-
   return (
     <div className="mt-10">
-      {newFiles.length > 0 && (
-        <div ref={recentRef}>
-          <div className="flex justify-between items-center">
-            <h2 className="text-lg font-medium">{t("newFiles")}</h2>
-            <button
-              onClick={() => {
-                recentRef.current?.remove();
-              }}
-              className="mr-2"
-            >
-              <X className="w-4 h-4" />
-            </button>
-          </div>
-          <hr className="mt-1 mb-5" />
-          <div className="grid grid-cols-[repeat(auto-fill,minmax(10rem,1fr))] sm:grid-cols-[repeat(auto-fill,16rem)] justify-items-start gap-3 sm:gap-2 mx-auto mb-3">
-            {newFiles.map((file) => (
-              <Fragment key={file.id}>
-                <ImagePreviewGrid
-                  file={file}
-                  selected={selected}
-                  onClick={() => {
-                    setStartIndex(files.indexOf(file));
-                    setCarouselOpen(true);
-                  }}
-                  onSelect={() => {}}
-                />
-              </Fragment>
-            ))}
-          </div>
-
-          <h2 className="text-lg font-medium mt-8">{t("albumContent")}</h2>
-          <hr className="mt-1 mb-5" />
-        </div>
-      )}
       {selecting ? (
         <SelectingBar
           selected={selected}
@@ -448,6 +403,12 @@ export const ImagesGrid = ({ sortState }: { sortState: ImagesSortMethod }) => {
       {folder.description ? (
         <FolderDescription className="block sm:hidden mb-3" />
       ) : null}
+      <RecentlyAdded
+        onClickImage={(file) => {
+          setStartIndex(files.indexOf(file));
+          setCarouselOpen(true);
+        }}
+      />
       <div
         className={cn(
           files.length === 0
